Add all serialization code to pipeline serializer
Remove the use of pipeline operator
