--- rocAL/source/pipeline/pipeline_operator.cpp
@@ -1,150 +1,0 @@
-/*
-Copyright (c) 2025 Advanced Micro Devices, Inc. All rights reserved.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-*/
-
-#include "pipeline/pipeline_operator.h"
-
-void set_tensor_proto(rocal_proto::InputOutput *in_out_proto, Tensor *tensor, bool is_input = false) {
-    in_out_proto->set_name(tensor->tensor_name());
-    in_out_proto->set_device(static_cast<int>(tensor->info().mem_type()));
-    in_out_proto->set_dtype(static_cast<int>(tensor->info().data_type()));
-    in_out_proto->set_layout(static_cast<int>(tensor->info().layout()));
-    in_out_proto->set_color_format(static_cast<int>(tensor->info().color_format()));
-    for (auto &dim : tensor->info().dims())
-        in_out_proto->add_dims(dim);
-    in_out_proto->set_num_dims(tensor->info().num_of_dims());
-    in_out_proto->set_is_argument_input(is_input);
-}
-
-void PipelineOperator::serialize_pipeop_inputs_and_outputs_to_protobuf(rocal_proto::OperatorDef *opdef) {
-    if (this->module_name == "reader")
-        return;  // Readers do not have tensor outputs, hence return
-
-    // Serialize input tensors to protobuffers
-    for (auto &node_input : this->node->input()) {
-        rocal_proto::InputOutput *input = opdef->add_inputs();
-        set_tensor_proto(input, node_input, true);
-    }
-
-    // Serialize output tensors to protobuffers
-    for (auto &node_output : this->node->output()) {
-        rocal_proto::InputOutput *output = opdef->add_outputs();
-        set_tensor_proto(output, node_output);
-    }
-}
-
-void PipelineOperator::serialize_pipeop_args_to_protobuf(rocal_proto::OperatorDef *opdef) {
-    std::vector<Argument> arguments_list;
-
-    if (this->module_name == "reader") {
-        arguments_list = this->arguments;
-    } else {
-        arguments_list = this->node->get_args_list();
-    }
-    // Iterate through each argument to store in the protobuffers
-    for (auto &op_arg : arguments_list) {
-        rocal_proto::Arguments *arg = opdef->add_args();
-        arg->set_name(op_arg.arg_name);
-        arg->set_type(op_arg.type_name);
-        arg->set_is_vector(op_arg.is_vector);
-
-        if (op_arg.type_name == "nullptr") continue;
-        if (op_arg.enum_type_name != "")
-            arg->set_instance_name(op_arg.enum_type_name);
-
-        if (op_arg.is_parameter) {
-            // TODO - Will be enabled later
-            // rocal_proto::Parameter *param = arg->mutable_param();
-            // serialize_parameter_to_protobuf(param, op_arg);
-        } else if (op_arg.type_name == "enum") {
-            rocal_proto::EnumType* enum_arg = arg->mutable_enum_value();
-            enum_arg->set_name(op_arg.enum_type_name);
-            enum_arg->set_value(std::any_cast<int>(op_arg.values[0]));
-        } else {
-            // Scalars go to the flat repeated fields; vectors go to repeated *Vector messages
-            if (op_arg.is_vector) {
-                if (op_arg.values.empty()) {
-                    // Represent empty vector by adding an empty vector message of the right type
-                    if (op_arg.type_name == "int" || op_arg.type_name == "shared_ptr"
-                        || op_arg.type_name == "unsigned" || op_arg.type_name == "size_t") {
-                        static_cast<void>(arg->add_int_vectors());
-                    } else if (op_arg.type_name == "float") {
-                        static_cast<void>(arg->add_float_vectors());
-                    } else if (op_arg.type_name == "char_str" || op_arg.type_name == "string"
-                               || op_arg.type_name == "map_string") {
-                        static_cast<void>(arg->add_string_vectors());
-                    } else {
-                        THROW("Vector type not supported for Argument " + op_arg.arg_name + " with type " + op_arg.type_name);
-                    }
-                } else {
-                    if (op_arg.type_name == "int" || op_arg.type_name == "unsigned" || op_arg.type_name == "size_t") {
-                        auto *vec = arg->add_int_vectors();
-                        for (auto &v : op_arg.values) {
-                            // Map unsigned/size_t to int64 for IntVector as per spec (only Int/Float/String vectors permitted)
-                            if (op_arg.type_name == "unsigned") {
-                                vec->add_values(static_cast<int64_t>(std::any_cast<unsigned>(v)));
-                            } else if (op_arg.type_name == "size_t") {
-                                vec->add_values(static_cast<int64_t>(std::any_cast<size_t>(v)));
-                            } else {
-                                vec->add_values(static_cast<int64_t>(std::any_cast<int>(v)));
-                            }
-                        }
-                    } else if (op_arg.type_name == "float") {
-                        auto *vec = arg->add_float_vectors();
-                        for (auto &v : op_arg.values) {
-                            vec->add_values(std::any_cast<float>(v));
-                        }
-                    } else if (op_arg.type_name == "char_str" || op_arg.type_name == "string" 
-                               || op_arg.type_name == "map_string") {
-                        auto *vec = arg->add_string_vectors();
-                        for (auto &v : op_arg.values) {
-                            vec->add_values(std::any_cast<std::string>(v));
-                        }
-                    } else {
-                        THROW("Vector type not supported for Argument " + op_arg.arg_name + " with type " + op_arg.type_name);
-                    }
-                }
-            } else {
-                // Scalar path (use flat repeated fields)
-                if (op_arg.values.size() > 1) {
-                    ERR("Argument has more than one value, is_vector should be set to true")
-                }
-                for (auto &v : op_arg.values) {
-                    if (op_arg.type_name == "int" || op_arg.type_name == "shared_ptr") {
-                        arg->add_ints(std::any_cast<int>(v));
-                    } else if (op_arg.type_name == "float") {
-                        arg->add_floats(std::any_cast<float>(v));
-                    } else if (op_arg.type_name == "char_str" || op_arg.type_name == "string") {
-                        arg->add_strings(std::any_cast<std::string>(v));
-                    } else if (op_arg.type_name == "bool") {
-                        arg->add_bools(std::any_cast<bool>(v));
-                    } else if (op_arg.type_name == "unsigned") {
-                        arg->add_uints(std::any_cast<unsigned>(v));
-                    } else if (op_arg.type_name == "size_t") {
-                        arg->add_uints(std::any_cast<size_t>(v));
-                    } else {
-                        THROW("Invalid type specified for the Argument " + op_arg.arg_name);
-                    }
-                }
-            }
-        }
-    }
-}
